Line break between Phase 1 and Phase 2
Two spaces at the end of a line (`  `) in Markdown indicates a hard break but not a new paragraph.  A simple hard break ends up being rendered as a space instead of a line break, so you can add the spaces to make the line break show up as well.  Added here.