Update RFQ.md
Updated RFQ to make the numbering align with the numbering in the eBuy solicitation.
Also addressed Issue #135 and includes updated Similar Experience language from the amendment posted to eBuy on 4/5/16.

Agile BPA vendors - please note that the eBuy solicitation is the official RFQ document of record.  18F and the Contracting Officers are working through a process to better sync between GH and eBuy for future RFQ's.